move diffusion case into test set
